+++ streams-bootstrap-core/src/testFixtures/java/com/bakdata/kafka/KafkaTest.java
@@ -48,6 +48,28 @@ public abstract class KafkaTest {
                         "PLAINTEXT://:9092,BROKER://:9093,CONTROLLER://:9094"); //TODO remove with 3.9.1 https://issues.apache.org/jira/browse/KAFKA-18281
     }
 
+    protected static void awaitProcessing(final ExecutableStreamsApp<?> app) {
+        awaitActive(app);
+        final ConsumerGroupVerifier verifier = ConsumerGroupVerifier.verify(app);
+        await()
+                .alias("Consumer group has finished processing")
+                .until(verifier::hasFinishedProcessing);
+    }
+
+    protected static void awaitActive(final ExecutableStreamsApp<?> app) {
+        final ConsumerGroupVerifier verifier = ConsumerGroupVerifier.verify(app);
+        await()
+                .alias("Consumer group is active")
+                .until(verifier::isActive);
+    }
+
+    protected static void awaitClosed(final ExecutableStreamsApp<?> app) {
+        final ConsumerGroupVerifier verifier = ConsumerGroupVerifier.verify(app);
+        await()
+                .alias("Consumer group is closed")
+                .until(verifier::isClosed);
+    }
+
     private static ConditionFactory await() {
         return Awaitility.await()
                 .pollInterval(Duration.ofSeconds(2L))
@@ -81,28 +103,6 @@ public abstract class KafkaTest {
 
     protected SchemaRegistryClient getSchemaRegistryClient() {
         return this.testTopologyFactory.getSchemaRegistryClient();
-    }
-
-    protected void awaitProcessing(final ExecutableStreamsApp<?> app) {
-        this.awaitActive(app);
-        final ConsumerGroupVerifier verifier = ConsumerGroupVerifier.verify(app);
-        await()
-                .alias("Consumer group has finished processing")
-                .until(verifier::hasFinishedProcessing);
-    }
-
-    protected void awaitActive(final ExecutableStreamsApp<?> app) {
-        final ConsumerGroupVerifier verifier = ConsumerGroupVerifier.verify(app);
-        await()
-                .alias("Consumer group is active")
-                .until(verifier::isActive);
-    }
-
-    protected void awaitClosed(final ExecutableStreamsApp<?> app) {
-        final ConsumerGroupVerifier verifier = ConsumerGroupVerifier.verify(app);
-        await()
-                .alias("Consumer group is closed")
-                .until(verifier::isClosed);
     }
 
 }