+++ streams-bootstrap-test/src/main/java/com/bakdata/kafka/TestTopologyFactory.java
@@ -84,17 +84,6 @@ public final class TestTopologyFactory {
     }
 
     /**
-     * Create {@code Configurator} to configure {@link org.apache.kafka.common.serialization.Serde} and
-     * {@link org.apache.kafka.common.serialization.Serializer} using the {@code TestTopology} properties.
-     * @param testTopology {@code TestTopology} to use properties of
-     * @return {@code Configurator}
-     * @see TestTopology#getProperties()
-     */
-    public static Configurator createConfigurator(final TestTopology<?, ?> testTopology) {
-        return new Configurator(testTopology.getProperties());
-    }
-
-    /**
      * Create a new Kafka Streams config suitable for test environments. This includes setting the following
      * parameters in addition to {@link #createStreamsTestConfig()}:
      * <ul>